+++ omf-core-framework/src/main/java/com/samares_engineering/omf/omf_core_framework/feature/registrables/rule_engines/rule/ARule.java
@@ -16,7 +16,7 @@ public abstract class ARule implements IRule<PropertyChangeEvent, PropertyChange
     public String id = "";
     public boolean isActivated = true;
 
-    public ARule(){};
+    public ARule(){}
 
     public ARule(String id){
         this.id = id;
@@ -29,7 +29,8 @@ public abstract class ARule implements IRule<PropertyChangeEvent, PropertyChange
 
     protected abstract boolean eventMatches(PropertyChangeEvent evt);
 
-    public abstract void debug(Object o);
+    @Deprecated
+    public void debug(Object o){}
 
     @Override
     public boolean isActivated() {